refactor (backend): renamed db_engine to db and update path import across all references
