+++ agenta-backend/agenta_backend/tests/variants_main_router/test_variant_testset_router.py
@@ -4,7 +4,7 @@ import httpx
 import pytest
 from sqlalchemy.future import select
 
-from agenta_backend.models.db_engine.shared import db_engine
+from agenta_backend.models.db.shared import db_engine
 from agenta_backend.models.db_models import (
     AppDB,
     TestSetDB,
