+++ agenta-backend/agenta_backend/tests/variants_main_router/test_variant_versioning_deployment.py
@@ -5,7 +5,7 @@ import random
 
 from sqlalchemy.future import select
 
-from agenta_backend.models.db_engine.shared import db_engine
+from agenta_backend.models.db.shared import db_engine
 from agenta_backend.models.db_models import (
     AppDB,
     TestSetDB,
